backend/requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
psycopg2==2.8.6
gunicorn==20.1.0
psutil==5.8.0
pyyaml==5.4.1
coverage==5.5
requests==2.25.1